64bit port
If this works on linux too, then we're golden.  intptr_t has the same length as void*, so on all systems, regardless of bit-length, all data will fit into this type.

+++ src/engine/doomtype.h
@@ -35,7 +35,7 @@
 typedef int             dboolean;
 typedef uint8_t         byte;
 typedef uint16_t	word;
-typedef unsigned long	dword;
+typedef intptr_t	dword;
 typedef int8_t          int8;
 typedef uint8_t         uint8;
 typedef int16_t         int16;
